This commit touches a lot of stuff.
Dates and now fully supported, no longer do we mash everything into LocalDateTimes. The temporals are even more abstract now. We'll see if we can take it a step further with timestamps.

UUIDs are supported, so we don't have to jam a bunch of strings around anymore. Still need much better support for generating java objects with unique constructors and initialization processes.

Closes #10.


Former-commit-id: c1f80299e41a07416de2b3dcd070dd2b33f529ed