This commit touches a lot of stuff.
Dates and now fully supported, no longer do we mash everything into LocalDateTimes. The temporals are even more abstract now. We'll see if we can take it a step further with timestamps.

UUIDs are supported, so we don't have to jam a bunch of strings around anymore. Still need much better support for generating java objects with unique constructors and initialization processes.

Closes #10.


Former-commit-id: 129640486c45bbf5b4c7cf62a756a04881187b07