Fixed a bunch of pom/depdendency issues. Here's what we have:
trestle-ontology bundles the oracle and virtuoso dependencies, because they don't exist on any public repos.
trestle-raster does the same thing.
trestle-actor creates an uberjar so we can run it from the command line, but we also create a dependency reduced pom, so it doesn't try to re-bundle the oracle jars. We also minimize the jar as much as possible.


Former-commit-id: 11379843a800d95717bedbaa5859890fce9562cb