TRESTLE-716: Re-enable javadoc building (#109)
* TRESTLE-716: Re-enable javadoc building

It turns out a JDK bug was causing Javadoc generation to fail, when the code source was set to Java 8. A simple configuration override solved the issue and it should go away when we move to Java 12+.

* TRESTLE-716: Fix javadoc failures

Cleaned up some failing javadocs, since we we're building them for a while they rotted away.