[Chore] Add support for building and running with Java 15 (#621)
* Upgrade Gradle to support Java 15

Upgrade Gradle to 6.7.1, which gives us support for building/running with Java 15.

We're still using 11 source and bytecode, but building with 15 is nice to keep us up to date with the tip of Java.

* Upgrade Azure and CI to JDK 15

Let's run on the latest.

Of course, we're also running on JDK 11 as well.

* Udpate Yaml file to try and make it work

Update Cache keys as well.

* Don't build with 15 in CI.