TRESTLE-757: Refactored everything into a list of notifications.
The notification components are now part of the Navigation module, which avoids having to inject and load the UIModule on app startup.

The NotificationCenter is now injected into the application by the NavBar component.
From there, it subscribes to the AppState and displays the current list of notifications.

Added a timeout callback to remove the notification from the store after 5s.