TRESTLE-733: Refactor TrestleObjectReader to make use of RxJava classes
Cleaned up the TrestleObjectReader class to make better use of RxJava features.

As part of this, greatly simplified some of the method signatures by moving away from the Optional and throw pattern that we've been (ab)using for a while now.

Rather than returning an optional that we immediately unpack into an exception, we just throw the exception and move on.