TRESTLE-733: Simplify logic for transaction handling
Add additional methods to track parent status of transactions, which makes it a lot simpler to debug and determine when we should actually return/abort transactions.

There are some additional improvements that should be made, we probably need to invert the inheritance of the transactions. So rather than sorting child -> parent, we should store parent -> children.