TRESTLE-733: Migrate Reasoner signatures to RxJava (#752)
* TRESTLE-733: Refactor ObjectEngineUtils

Refactor the getAdjustedTemporal method to return a Single.
This moves the blocking get up another level and helps simplify the code.

Also added some unit tests which make use of a mock Ontology, to help speed up testing.

* TRESTLE-733: Refactor TrestleObjectReader to make use of RxJava classes

Cleaned up the TrestleObjectReader class to make better use of RxJava features.

As part of this, greatly simplified some of the method signatures by moving away from the Optional and throw pattern that we've been (ab)using for a while now.

Rather than returning an optional that we immediately unpack into an exception, we just throw the exception and move on.

* TRESTLE-733: Refactor TrestleObjectWriter to use RxJava.

Initial refactor of TrestleObjectWriter.
It's a pretty terrible change at this point, lots of breaking tests, but hopefully we'll get it settled down.

* TRESTLE-733: Finished migrating TrestleObjectWriter to RxJava

Fun fact, Completable.defer() is the method we want, not `fromCallable()`.

Lots of refactoring here, tried to split some things into their own methods.
Working with `Completable` is tricky, because ignoring the result causes lots of things to fail.
We'll need a lot of manual touch-up.

* TRESTLE-733: Update downstream modules to support new RxJava signatures

The only changes made were to get things compiling, we still need to actually fix the underlying logic.

* TRESTLE-733: Initial refactor of CollectionEngine

Object removal doesn't work correctly. Ugh.

* TRESTLE-733: TrestleEventEngine now support RxJava

Migrated the Event Engine methods to RxJava and cleaned up some of the calling code.

* TRESTLE-733 DataExportEngine is RxJava now

There are some really gross uses of raw generics, which should be addressed at some point.

* TRESTLE-733: SpatialEngine is now using RxJava

* Update Log4J configuration.

As in, make it actually work.

We're now including Log4J as one of our test dependencies, which makes it pretty easy to setup certain configuration properties.

* TRESTLE-733: Migrate SpatialEngine to RxJava

It's broken, quite broken though.

* TRESTLE-733: IndividualEngine now uses RxJava

* TRESTLE-733: Rewrite AggregationEngine logic

RxJava, and a whole lot of other bugs.

* TRESTLE-733: Improve threading of TrestleTransactions

Thread transactions between method calls in a more structured manner.

* TRESTLE-733: Simplify logic for transaction handling

Add additional methods to track parent status of transactions, which makes it a lot simpler to debug and determine when we should actually return/abort transactions.

There are some additional improvements that should be made, we probably need to invert the inheritance of the transactions. So rather than sorting child -> parent, we should store parent -> children.

* TRESTLE-733: SpatialEngine tests are passing once again.

Dramatically improved threading and transaction handling amongst all the various helper methods.

* TRESTLE-733: Refactor spatial/temporal relationship calculator into its own class.

* TRESTLE-733: Spatial and Temporal relationships are working in their refactored state

* Everything except for the EqualityEngine

* TRESTLE-733: EqualityEngine migrated to RxJava

Created a new bug that we're tacking at TRESTLE-763.

* TRESTLE-733: Server and Integrations now compile with RxJava.

* TRESTLE-733: Address Codacy warnings

PR check cleanups.

* TRESTLE-733: Add tests for individual removal

A bit of refactoring by moving TrestleMissingIndividualException into the reasoner module, instead of common.

* TRESTLE-733: Individual removal methods are now reactive

Reworked the remove methods to make full use of RxJava.

* TRESTLE-763: Rollback aborted transactions

Fixed abort handling in the TransactingOntology class, to make what we already have for committed transactions.

* TRESTLE-733: Fix Javadocs

Update javadocs with new RxJava signatures.

* TRESTLE-733: More PR cleanups and fix incorrect test

A few cleanups for Codacy.

Fixed the individual removal test, it was failing but passing incorrectly.
It now fails correctly, which passes.

* TRESTLE-733: Fix throws doc

* TRESTLE-733: Reduce logging

* TRESTLE-733: Fix logging conflicts with server

Updated gradle files to only include the log4j testing jars with the reasoner.
We'll probably need to go back and fix this up for other modules at some point.